requirements.txt
asgiref==3.8.1
attrs==25.3.0
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.2
cssselect==1.3.0
defusedxml==0.7.1
Django==4.2
filelock==3.18.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==5.3.1
packaging==24.2
parsel==1.10.0
Protego==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==25.0.0
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
Scrapy==2.12.0
service-identity==24.2.0
setuptools==77.0.3
sqlparse==0.5.3
tldextract==5.1.3
Twisted==24.11.0
typing_extensions==4.12.2
urllib3==2.3.0
w3lib==2.3.1
zope.interface==7.2
mysql_connector_python==9.2.0
